+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v2.2.3 include:
+- Switch autogain algorithm to readsb builtin
 
 > [!NOTE]
 > Based on the available usage information, I have significantly reduced the number of images provided here. If there's one that you need for a different SBC which is supported either by Armbian or DietPi, please post a request on the [Zulip server](https://adsblol.zulipchat.com/#narrow/stream/391168-adsb-feeder-image)